--- tests/basic.expect
@@ -1,1 +1,0 @@
-Success!
