+++ tests/basic/CMakeLists.txt
@@ -1,0 +1,21 @@
+cmake_minimum_required(VERSION 3.21)
+include($ENV{XMOS_CMAKE_PATH}/xcommon.cmake)
+project(basic)
+
+# root
+set(XMOS_SANDBOX_DIR ${CMAKE_SOURCE_DIR}/../../..)
+
+# target: choose the target platform
+set(APP_HW_TARGET XCORE-AI-EXPLORER)
+
+# Includes and flags
+set(APP_INCLUDES src)
+set(APP_COMPILER_FLAGS
+  -O2
+  -g
+  -report
+  -Wall
+)
+
+set(APP_DEPENDENT_MODULES lib_device_control)
+XMOS_REGISTER_APP()
