--- tests/basic/Makefile
@@ -1,5 +1,0 @@
-TARGET = XCORE-200-EXPLORER
-USED_MODULES = lib_device_control
-XCC_FLAGS = -g
-XMOS_MAKE_PATH ?= ../..
--include $(XMOS_MAKE_PATH)/xcommon/module_xcommon/build/Makefile.common
