--- tests/basic/wscript
@@ -1,12 +1,0 @@
-def configure(conf):
-    conf.load('xwaf.compiler_xcc')
-
-
-def build(bld):
-    bld.env.TARGET_ARCH = 'XCORE-200-EXPLORER'
-    bld.env.XCC_FLAGS = ['-g']
-
-    # Build our program
-    prog = bld.program(
-        source='src/test.xc',
-        depends_on='lib_device_control')
