+++ tests/utils.py
@@ -2,6 +2,7 @@
 # Copyright 2024 XMOS LIMITED.
 # This Software is subject to the terms of the XMOS Public Licence: Version 1.
 import subprocess
+from pathlib import Path
 
 def xsim_firmware(xe_file, check_return_code=True, return_output=True, timeout_s=60):
     """
@@ -18,6 +19,7 @@ def xsim_firmware(xe_file, check_return_code=True, return_output=True, timeout_s
     None: If `return_output` is False.
 
     Raises:
+    FileNotFoundError: If the .xe file does not exist.
     subprocess.CalledProcessError: If `check_return_code` is True and the command exits with a non-zero status.
     subprocess.TimeoutExpired: If the command times out.
 
@@ -25,7 +27,12 @@ def xsim_firmware(xe_file, check_return_code=True, return_output=True, timeout_s
     >>> output = xsim_firmware("path/to/firmware.xe")
     >>> print(output)
     """
-    cmd = f"xsim {xe_file}"
+
+    xe_path = Path(xe_file)
+    if not xe_path.is_file():
+        raise FileNotFoundError(f"The file {xe_file} does not exist.")
+
+    cmd = f"xsim {xe_path}"
     print(f"Running: {cmd}")
     output = None
     try:
