--- tests/version.expect
@@ -1,1 +1,0 @@
-Success!
